Solved minor bugs and added dump of console to file log.txt

+++ JAGEgradle/JAGEmodules/src/main/java/com/valarion/gameengine/events/Player.java
@@ -334,7 +334,7 @@ public class Player implements Event, Serializable {
 	@Override
 	public void render(GameContainer container, Graphics g, int tilewidth, int tileheight) throws SlickException {
 		if (sprite != null) {
-			sprite.draw(getXDraw(tilewidth), getYDraw(tileheight));
+			sprite.draw((int)getXDraw(tilewidth), (int)getYDraw(tileheight));
 		}
 	}
